fix chal
