+++ src/app/(mobile)/challenge/grammar/(quiz)/korean/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import RandomKoreanGrammarQuiz from "@/components/challenge/RandomKoreanGrammarQuiz";
-import { useSearchParams } from "next/navigation";
+import { useUser } from "@/hooks/useUser";
 import { Suspense } from "react";
 
 const KoreanGrammarQuizPage = () => {
@@ -13,8 +13,8 @@ const KoreanGrammarQuizPage = () => {
 };
 
 const KoreanGrammarQuiz = () => {
-  const searchParams = useSearchParams();
-  const userId = searchParams?.get("userId");
+  const { userInfo } = useUser();
+  const userId = userInfo?.id;
 
   return <div>{userId && <RandomKoreanGrammarQuiz userId={userId} />}</div>;
 };
