+++ src/app/(mobile)/challenge/grammar/result/page.tsx
@@ -5,10 +5,10 @@ import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 import { Suspense } from "react";
 
-import image0 from "@//assets/result/result-0.svg";
-import image1 from "@//assets/result/result-1.svg";
-import image2 from "@/assets/result/result-2.svg";
-import image3 from "@/assets/result/result-3.svg";
+import zeroPoint from "@//assets/result/zero-point.svg";
+import onePoint from "@//assets/result/one-point.svg";
+import twoPoint from "@/assets/result/two-point.svg";
+import threePoint from "@/assets/result/three-point.svg";
 
 const ResultPage = () => {
   return (
@@ -25,10 +25,10 @@ const Result = () => {
   return (
     <div className="w-full flex flex-col relative min-h-screen gap-4">
       <div className="w-full flex flex-col justify-center min-h-[calc(100vh-150px)]">
-        {message === "점수: 0 / 3" ? <Image src={image0} alt="Result 1" width={375} height={550} /> : null}
-        {message === "점수: 1 / 3" ? <Image src={image1} alt="Result 1" width={375} height={550} /> : null}
-        {message === "점수: 2 / 3" ? <Image src={image2} alt="Result 2" width={375} height={550} /> : null}
-        {message === "점수: 3 / 3" ? <Image src={image3} alt="Result 3" width={375} height={550} /> : null}
+        {message === "점수: 0 / 3" ? <Image src={zeroPoint} alt="zero" width={375} height={550} /> : null}
+        {message === "점수: 1 / 3" ? <Image src={onePoint} alt="one" width={375} height={550} /> : null}
+        {message === "점수: 2 / 3" ? <Image src={twoPoint} alt="two" width={375} height={550} /> : null}
+        {message === "점수: 3 / 3" ? <Image src={threePoint} alt="three" width={375} height={550} /> : null}
       </div>
       <div className="w-full absolute bottom-[31px]">
         <div className="flex flex-col gap-[10px]">
